pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
